Remove global traps
This is what I tried so that rspec-core can play well with REPLs &
debuggers.

The problem is that rspec sets a global trap for the Ctrl-C
signal that exits rspec. If pry or byebug are running, this is most
likely not intended by the user.

The solution I thought of is to remove the global trap and rescue
interrupt instead. It seemd to me the only possible solution because
global traps for signals can't be disabled nor overriden.